voila/requirements.txt
Click==7.0
Cython==0.29.14
Flask==1.0.2
Flask-Session==0.3.1
Flask-WTF==0.14.2
gitdb2==2.0.6
GitPython==3.0.5
gunicorn==19.9.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
psutil==5.6.7
scipy==1.4.1
six==1.14.0
smmap2==2.0.5
waitress==1.1.0
Werkzeug==0.16.0
WTForms==2.2.1